some refactors + add "N/A" difficulty class
