+++ content/6_Advanced/FFT-ext.problems.json
@@ -2,11 +2,11 @@
   "MODULE_ID": "fft-ext",
   "ys": [
     {
-      "uniqueId": "ys-PartitionFunction!!",
+      "uniqueId": "ys-PartitionFunction",
       "name": "Partition Function",
       "url": "https://judge.yosupo.jp/problem/partition_function",
       "source": "YS",
-      "difficulty": "?",
+      "difficulty": "N/A",
       "isStarred": false,
       "tags": [],
       "solutionMetadata": {
@@ -18,7 +18,7 @@
       "name": "Bernoulli Number",
       "url": "https://judge.yosupo.jp/problem/bernoulli_number",
       "source": "YS",
-      "difficulty": "?",
+      "difficulty": "N/A",
       "isStarred": false,
       "tags": [],
       "solutionMetadata": {
